Add files via upload
Updated files form V1.1 to V1.2, added new options and solved problem with openvpn client connection fails
If have older versions, clone project again, and do upgrade.sh instead of Instalación.sh

+++ update.sh
@@ -2,7 +2,7 @@
 
 ##### ACTUALIZADOR RPI-NAVARRUX - JL NAVARRO ADAM - LICENCIA CC
 
-#### Version 1.1
+#### Version 1.2
 
 clear
 
@@ -17,7 +17,9 @@ tar -xvzf rpi-navarrux.tgz
 cp -ar /tmp/rpi-navarrux/etc/rc.local /etc/
 cp -ar /tmp/rpi-navarrux/root/MENU /root/
 cp -ar /tmp/rpi-navarrux/root/bin/* /root/bin/
-cp -ar /tmp/rpi-navarrux/root/conf/inadyn.conf /root/conf/
+if [ ! -f /root/conf/inadyn.conf ]; then
+    cp -ar /tmp/rpi-navarrux/root/conf/inadyn.conf /root/conf/
+fi
 # ELIMINAMOS TEMPORALES
 rm -R /tmp/rpi-navarrux
 echo ""
